+++ src/hal/impl/rp2040/hal_serial.cpp
@@ -5,7 +5,7 @@
 #include "../../hal_sync.h"
 #include "../../hal_system.h"
 #include "../shared/hal_mutex_once.h"
-#include "hal/impl/shared/compat/debug_format/hal_debug_format.h"
+#include "hal/impl/shared/debug/hal_debug_format.h"
 #include <stdarg.h>
 #include <stdio.h>
 #include <string.h>